Add files via upload
These files are associated with the DAMSystem Software. There are also instructions under "DAMSystem3 Software Data Sheet" that will guide an individual on how to install the DAMSystem3 Data Collection Software. The other excel sheets are used to then organize the raw data obtained and organize the data into a readable excel spreadsheet. The October 26 excel sheets are examples for what the data should come out to before you can use the code. "Oct26_aaaflyanalysis_1_minute.xls" is the data received after the DAM File system has organized it. Then a new excel sheet was formed that could be input into the code and used to make graphs and run ANOVA tests.